Renaming package to be AstrodbKit
