+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = astrodbkit2
+name = astrodbkit
 author = David Rodriguez
 edit_on_github = False
-github_project = dr-rodriguez/AstrodbKit2
+github_project = astrodbtoolkit/AstrodbKit
 # Most details are now in pyproject.toml
